Added A video tutorial I posted on flare.pub
Hi guys. I'm writing a bitcoin educational guide with video tutorials. Anyway, I think Current is a great way to onboard pre-coiners so I did a Current Tutorial on flare.pub.  The video is hosted on nostr. build. that's probably not the best way to host it, but let me know what you think. Thanks guys. I love the project!